Changed lightmap calculation to using G3AccumVertColor().
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@1477 01821946-851b-0410-b7f4-f1477c70e57a

+++ gameio/gamemine.cpp
@@ -1011,7 +1011,7 @@ for (segP = gameData.segs.segments + i; i < j; i++, segP++) {
 		m = (pl->nSegment < 0) ? gameData.objs.objects [pl->nObject].nSegment : pl->nSegment;
 		if (!SEGVIS (m, i))
 			continue;
-		h = VmVecDist (&center, &pl->vPos) - fl2f (pl->rad) / 10.0f;
+		h = (int) (VmVecDist (&center, &pl->vPos) - fl2f (pl->rad) / 10.0f);
 		if (h > MAX_LIGHT_RANGE * pl->range)
 			continue;
 		pDists [n].nDist = h;
